[c++][fix] resolve undefined reference errors for DLPack
